+++ R/factory.R
@@ -162,6 +162,15 @@ factory <- function(tab,
   }
   align <- strsplit(align, "")[[1]]
 
+  # dot align with unicode spaces (latex has its own mechanism)
+  if (!settings_equal("output_format", "latex")) {
+    align_d <- grep("d", align)
+    for (i in align_d) {
+      tab[[i]] <- pad(tab[[i]], style = "character")
+    }
+    align[align == "d"] <- "c"
+  }
+
   ## build table
   out <- factory_fun(tab,
     align = align,