+++ tests/testthat/test-mathmode.R
@@ -2,23 +2,9 @@ mod <- list(
     lm(mpg ~ hp, mtcars),
     lm(mpg ~ hp + drat, mtcars))
 
-test_that("d-column: supported outputs", {
-    expect_error(modelsummary(mod, output = "latex", align = "ldd"), NA)
-    expect_error(modelsummary(mod, output = "latex_tabular", align = "ldd"), NA)
-    expect_error(modelsummary(mod, output = "html", align = "ldd"), "only supported")
-    expect_error(modelsummary(mod, output = "kableExtra", align = "ldd"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "gt", align = "ldd"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "huxtable", align = "ldd"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "flextable", align = "ldd"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "markdown", align = "ldd"), regexp = "only supported")
-})
-
-
-
 test_that("d-column: known output", {
     expect_snapshot(modelsummary(mod, align = "ldd", output = "latex"))
 })
-
 
 
 test_that("HTML global options", {